KlipperPatch
I love this plugin because it ties in seamlessly with material profiles. However, having recently switched to Klipper to test it out, I was sad that this plugin didn't work with Klipper's Pressure Advance feature. So I created a Klipper macro that makes this plugin work flawlessly. 

All you have to do is add the above macro to your Klipper configuration file(printer.cfg), restart the firmware(FIRMWARE_RESTART) and then M900 K works to update Pressure Advance with the value you enter in the Linear Advance field.